вносит правки по ревью
